Introduce DisplaceManager class.
